requirements.txt
pymodbus
paho-mqtt
pyserial
flask
gunicorn
requests
psycopg2-binary
pyyaml